refactor(react-query-devtools): move QueryStatusCount header to subdirectory
